backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.3.3
python-multipart==0.0.6
